msg blade

+++ garage/app/Http/Controllers/MechanicController.php
@@ -88,6 +88,9 @@ class MechanicController extends Controller
      */
     public function destroy(Mechanic $mechanic)
     {
+        if($mechanic->getTrucks()->count()) {
+            return redirect()->back()->with('info_msg', 'Turi atidirbti dvi savaites.');
+        }
         $mechanic->delete();
         return redirect()->route('m_index');
     }
